Move location of multi-cancellation-token
This avoids a code-scanning warning.
