+++ extensions/ql-vscode/src/common/vscode/multi-cancellation-token.ts
